Removed "com" folder. Separed domain from the spring implementation. Creating some space for other tech imps
