--- src/main/java/com/engman/EngmanApplication.java
@@ -1,14 +1,0 @@
-package com.engman;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-
-@SpringBootApplication
-public class EngmanApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EngmanApplication.class, args);
-	}
-}
