--- src/main/java/com/engman/common/utilities/mappers/ModelMapperConfig.java
@@ -1,13 +1,0 @@
-package com.engman.common.utilities.mappers;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ModelMapperConfig {
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-}
