--- src/main/java/com/engman/common/utilities/mappers/ModelMapperService.java
@@ -1,7 +1,0 @@
-package com.engman.common.utilities.mappers;
-
-import org.modelmapper.ModelMapper;
-
-public interface ModelMapperService {
-    ModelMapper forResponse();
-    ModelMapper forRequest();}
