--- src/main/java/com/engman/domain/balance/ITransferrable.java
@@ -1,15 +1,0 @@
-package com.engman.domain.balance;
-
-public interface ITransferrable {
-
-    //On every iteration/hit/scale
-    Integer Transfer(Integer bit);
-    Integer GetCurrent();
-}
-
-/*
-Project     -> Scope+++
-Budget      -> Budget--
-Developer   -> Revenue--
-Skill       ->
- */
