--- src/main/java/com/engman/entities/abstracts/MainRepository.java
@@ -1,8 +1,0 @@
-package com.engman.entities.abstracts;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.repository.NoRepositoryBean;
-
-@NoRepositoryBean
-public interface MainRepository<T, ID>  extends MongoRepository<T, ID> {
-}
