--- src/main/java/com/engman/entities/abstracts/SkillRepository.java
@@ -1,6 +1,0 @@
-package com.engman.entities.abstracts;
-
-import com.engman.entities.concretes.Skill;
-
-public interface SkillRepository extends MainRepository<Skill, Integer> {
-}
