--- src/main/java/com/engman/services/abstracts/SkillService.java
@@ -1,11 +1,0 @@
-package com.engman.services.abstracts;
-
-import com.engman.services.requests.CreateSkillRequest;
-import com.engman.services.responses.GetAllSkillsResponse;
-
-import java.util.List;
-
-public interface SkillService {
-    void addSkill(CreateSkillRequest createSkillRequest);
-    List<GetAllSkillsResponse> getSkills();
-}
