+++ src/main/java/engman/domain/balance/ITransferrable.java
@@ -1,0 +1,15 @@
+package engman.domain.balance;
+
+public interface ITransferrable {
+
+    //On every iteration/hit/scale
+    Integer Transfer(Integer bit);
+    Integer GetCurrent();
+}
+
+/*
+Project     -> Scope+++
+Budget      -> Budget--
+Developer   -> Revenue--
+Skill       ->
+ */
