+++ src/main/java/engman/spring/EngmanApplication.java
@@ -1,0 +1,12 @@
+package engman.spring;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EngmanApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EngmanApplication.class, args);
+	}
+}
