+++ src/main/java/engman/spring/common/utilities/mappers/ModelMapperConfig.java
@@ -1,0 +1,13 @@
+package engman.spring.common.utilities.mappers;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ModelMapperConfig {
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+}
