+++ src/main/java/engman/spring/common/utilities/mappers/ModelMapperManager.java
@@ -1,0 +1,27 @@
+package engman.spring.common.utilities.mappers;
+
+import lombok.AllArgsConstructor;
+import org.modelmapper.ModelMapper;
+import org.modelmapper.convention.MatchingStrategies;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class ModelMapperManager implements ModelMapperService{
+    private ModelMapper modelMapper;
+    @Override
+    public ModelMapper forResponse() {
+        this.modelMapper.getConfiguration()
+                .setAmbiguityIgnored(true)
+                .setMatchingStrategy(MatchingStrategies.LOOSE);
+        return this.modelMapper;
+    }
+
+    @Override
+    public ModelMapper forRequest() {
+        this.modelMapper.getConfiguration()
+                .setAmbiguityIgnored(true)
+                .setMatchingStrategy(MatchingStrategies.STRICT);
+        return this.modelMapper;
+    }
+}
