+++ src/main/java/engman/spring/common/utilities/mappers/ModelMapperService.java
@@ -1,0 +1,7 @@
+package engman.spring.common.utilities.mappers;
+
+import org.modelmapper.ModelMapper;
+
+public interface ModelMapperService {
+    ModelMapper forResponse();
+    ModelMapper forRequest();}
