+++ src/main/java/engman/spring/entities/abstracts/MainRepository.java
@@ -1,0 +1,8 @@
+package engman.spring.entities.abstracts;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.repository.NoRepositoryBean;
+
+@NoRepositoryBean
+public interface MainRepository<T, ID>  extends MongoRepository<T, ID> {
+}
