+++ src/main/java/engman/spring/entities/abstracts/SkillRepository.java
@@ -1,0 +1,6 @@
+package engman.spring.entities.abstracts;
+
+import engman.spring.entities.concretes.Skill;
+
+public interface SkillRepository extends MainRepository<Skill, Integer> {
+}
