+++ src/main/java/engman/spring/services/abstracts/SkillService.java
@@ -1,0 +1,11 @@
+package engman.spring.services.abstracts;
+
+import engman.spring.services.requests.CreateSkillRequest;
+import engman.spring.services.responses.GetAllSkillsResponse;
+
+import java.util.List;
+
+public interface SkillService {
+    void addSkill(CreateSkillRequest createSkillRequest);
+    List<GetAllSkillsResponse> getSkills();
+}
