+++ src/main/java/engman/spring/services/requests/CreateSkillRequest.java
@@ -1,0 +1,14 @@
+package engman.spring.services.requests;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class CreateSkillRequest {
+    private Integer id;
+    private String name;
+    private String desc;
+}
