+++ src/main/java/engman/spring/services/responses/GetAllSkillsResponse.java
@@ -1,0 +1,15 @@
+package engman.spring.services.responses;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class GetAllSkillsResponse {
+    private Integer id;
+    private String name;
+    private String desc;
+    private String fourthfield;
+}
