--- src/test/java/com/engman/engman/EngmanApplicationTests.java
@@ -1,13 +1,0 @@
-package com.engman.engman;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EngmanApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
