+++ src/test/java/engman/spring/engman/EngmanApplicationTests.java
@@ -1,0 +1,13 @@
+package engman.spring.engman;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EngmanApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
